test: organize TypeInfo workflow fixtures
